get wealth moved
