+++ tests/conftest.py
@@ -52,7 +52,7 @@ import gym
 import pytest
 from aea_ledger_cosmos import CosmosCrypto
 from aea_ledger_ethereum import EthereumCrypto
-from aea_ledger_fetchai import FetchAICrypto
+from aea_ledger_fetchai import FetchAICrypto, FetchAIFaucetApi
 from cosmpy.clients.signing_cosmwasm_client import SigningCosmWasmClient
 from cosmpy.common.rest_client import RestClient
 from cosmpy.crypto.address import Address as CosmpyAddress
@@ -84,7 +84,7 @@ from aea.crypto.ledger_apis import (
     ETHEREUM_DEFAULT_CURRENCY_DENOM,
     FETCHAI_DEFAULT_ADDRESS,
 )
-from aea.crypto.registries import ledger_apis_registry, make_crypto
+from aea.crypto.registries import ledger_apis_registry, make_crypto, make_ledger_api
 from aea.crypto.wallet import CryptoStore
 from aea.exceptions import enforce
 from aea.helpers.base import CertRequest, SimpleId, cd
@@ -1482,3 +1482,23 @@ def method_scope(cls):
     cls.setup = lambda self: old_setup_class()
     cls.teardown = lambda self: old_teardown_class()
     return cls
+
+
+def get_wealth_if_needed(address: Address):
+    """
+     Get wealth from fetch.ai faucet to specific address
+
+    :param: address: Addresse to be funded from faucet
+    """
+    fetchai_api = make_ledger_api(FetchAICrypto.identifier, **FETCHAI_TESTNET_CONFIG)
+
+    balance = fetchai_api.get_balance(address)
+    if balance == 0:
+        FetchAIFaucetApi().get_wealth(address)
+
+        timeout = 0
+        while timeout < 40 and balance == 0:
+            time.sleep(1)
+            timeout += 1
+            _balance = fetchai_api.get_balance(address)
+            balance = _balance if _balance is not None else 0
